small fixes on setup.php and a bring update.php to work
