+++ CMakeLists.txt
@@ -111,7 +111,6 @@ set(CUSTOMFILES
     utils/query.php
     utils/server_compare.php
     utils/setup.php
-    utils/setupClass.php
     utils/specialphrases.php
     utils/update.php
     utils/warm.php
