+++ utils/setup.php
@@ -1,22 +1,21 @@
 #!@PHP_BIN@ -Cq
 <?php
 
-require_once(dirname(dirname(__FILE__)).'/settings/settings.php');
-require_once(CONST_BasePath.'/lib/init-cmd.php');
-include_once(CONST_InstallPath.'/utils/setupClass.php');
-// ->indirect via init-cmd.php -> /lib/cmd.php                  for runWithEnv, getCmdOpt
-// ->indirect via init-cmd.php -> /lib/init.php -> db.php       for &getDB()
+require_once dirname(dirname(__FILE__)) . '/settings/settings.php';
+require_once CONST_BasePath . '/lib/init-cmd.php';
+require_once CONST_BasePath . '/lib/SetupClass.php';
+// ->indirect via init-cmd.php->/lib/cmd.php                  for runWithEnv, getCmdOpt
+// ->indirect via init-cmd.php->/lib/init.php->db.php       for &getDB()
 
-include_once(CONST_BasePath.'/lib/setup_functions.php');
-include_once(CONST_BasePath.'/lib/setup_functions.php');
+require_once CONST_BasePath . '/lib/setup_functions.php';
 ini_set('memory_limit', '800M');
 
+use Nominatim\Setup\SetupFunctions as SetupFunctions;
 
 $aCMDOptions = createSetupArgvArray();
 
 // $aCMDOptions passed to getCmdOpt by reference
 getCmdOpt($_SERVER['argv'], $aCMDOptions, $aCMDResult, true, true);
-
 
 $bDidSomething = false;
 
@@ -38,16 +37,17 @@ if ($aCMDResult['osmosis-init']) {
 // ******************************************************
 // instantiate Setup class
 $cSetup = new SetupFunctions($aCMDResult);
-if ($aCMDResult['create-db'] || $aCMDResult['all']) {
-    $bDidSomething = true;
-    $cSetup -> createDB();
-}
 
 // *******************************************************
 // go through complete process if 'all' is selected or start selected functions
+if ($aCMDResult['create-db'] || $aCMDResult['all']) {
+    $bDidSomething = true;
+    $cSetup->createDB();
+}
+
 if ($aCMDResult['setup-db'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> setupDB();
+    $cSetup->setupDB();
 }
 
 // Try accessing the C module, so we know early if something is wrong
@@ -57,68 +57,68 @@ if (!checkModulePresence()) {
 
 if ($aCMDResult['import-data'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> importData($aCMDResult['osm-file']);
+    $cSetup->importData($aCMDResult['osm-file']);
 }
- 
+
 if ($aCMDResult['create-functions'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> createFunctions();
+    $cSetup->createFunctions();
 }
 
 if ($aCMDResult['create-tables'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> createTables();
-    $cSetup -> recreateFunction();
+    $cSetup->createTables();
+    $cSetup->createFunctions();
 }
 
 if ($aCMDResult['create-partition-tables'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> createPartitionTables();
+    $cSetup->createPartitionTables();
 }
 
 if ($aCMDResult['create-partition-functions'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> createPartitionFunctions();
+    $cSetup->createPartitionFunctions();
 }
 
 if ($aCMDResult['import-wikipedia-articles'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> importWikipediaArticles();
+    $cSetup->importWikipediaArticles();
 }
 
 if ($aCMDResult['load-data'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> loadData($aCMDResult['disable-token-precalc']);
+    $cSetup->loadData($aCMDResult['disable-token-precalc']);
 }
-    
+
 if ($aCMDResult['import-tiger-data']) {
     $bDidSomething = true;
-    $cSetup -> importTigerData();
+    $cSetup->importTigerData();
 }
 
 if ($aCMDResult['calculate-postcodes'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> calculatePostcodes($aCMDResult['all']);
+    $cSetup->calculatePostcodes($aCMDResult['all']);
 }
 
 if ($aCMDResult['index'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> index($aCMDResult['index-noanalyse']);
+    $cSetup->index($aCMDResult['index-noanalyse']);
 }
 
 if ($aCMDResult['create-search-indices'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> createSearchIndices();
+    $cSetup->createSearchIndices();
 }
 
 if ($aCMDResult['create-country-names'] || $aCMDResult['all']) {
     $bDidSomething = true;
-    $cSetup -> createCountryNames($aCMDResult);
+    $cSetup->createCountryNames($aCMDResult);
 }
 
 if ($aCMDResult['drop']) {
     $bDidSomething = true;
-    $cSetup -> drop($aCMDResult);
+    $cSetup->drop($aCMDResult);
 }
 
 // ******************************************************
